requirements.txt
streamlit
pandas
python-dotenv
PlexAPI
requests
